Added remaining changes from 17-06-25 and some other bugfixes
Added remaining changes from 17-06-25

-------

- Created BenchSet prefab. A set of bench with 3 rows.
- Created a new StaticCam, which shows the whole field. This is the 4th camera in the camera switcher.

- Change the CameraFollow Script. Now the name is 'CubeCamera', and I added some getters and setters.

- FIXED: Bug where if any of the cube moved, the camera would'nt follow them. Now it sets the current following object position when the camera is active.
- FIXED: If the camera was set off and the offset position wasn't calculated, the offset position was set at the moment that camera got activated again. In order to solve this, the offset position has to be a fixed number set from the beginning.

--

BUGS FOUND: The movement isnt aligned with the direction of Kai, so it appears that its moving sides ways. A possible fix to this problem is just to rotate the Visual to match the movement, but its probably better to have a direction vector in the script or something like that.

Possible changes: Is kinda weird that one control scheme is called "WASD_ARROWS" and the other "IJKL_SHIFT". Maybe is just me tho, but could be good to change the ijkl name in the control scheme.