New skin selection UI (integrated into main menu UI),2 skins added and 2v2 skin support!
I have moved the old skin select scene into a "Old or Unused" folder in Assets/Scenes. It has now been replaced with the new UI system for selection of skin which is integrated into the main menu.

I have added 2 skins, Tyler and Wai, along with short details for each character.

Skins are now affected in the 2v2 mode as well, which now allows you to use your favorite characters in both training mode AND the field mode.